Refs #24078 : LoadInstrument also from nexus files
In LoadEventNexus, after having checked if an instrument XML is found in the file, first check if the file contains geometry in
the binary nexus data before searching for the instrument name.
The file is checked using the LoadGeometry::isNexus() function and if it returns true, LoadInstrument is called with a
filename as argument rather than instrument name.