Avoid empty mount point overriding replacement path
When a replacement is done by ICatalogInfo::transformArchivePath, the replacement prefix is being overridden by the preferred mount point based on whether an Optional is set. However, if the user property is not provided this was being set to an empty string. This commit sets it to an uninitialised optional instead.