Change rebinning in Stitch1DMany
The rebinning step will be used for intermediate rebinning.
The rebinning boundaries will be used for final rebinning.
Now, start and end overlap values are sensible for stitching multiple ws

OTHERWISE THE BINNING AFTER STITCHING THE FIRST TWO WORKSPACES DEFINES AN END OVERLAP EQUAL TO THE END BIN FOR THE NEXT STITCH AND SO ON.

- binIndexOf handling: min and max index known and now default values

Refs #22921