Explicitely use ln (not log) in parameter file
This is used for the Parallax algorithm which uses muParser.
muParser has changed the meaning of log in version 2.2.5 from log10 to ln (Thanks for finding this out, MartynGigg).
The build servers do not provide the same version of the muParser.
To prevent future confusion, we like to use ln instead of log.

The existing iq D22 reference workspace is valid and should be used

Refs #24364