Remove ancient custom FindGSL cmake module
It caused an additional linker flag, -L/usr/lib/x86_64-linux-gnu, to
be added where ever the GSL_LIBRARIES variable was used. On Ubuntu 18.04
this seems to cause the linker to pick libomp over libgomp and when a
library built with GSL_LIBRARIES is then linked to an executable
built without it we get the executable link to both libomp and libgomp
resulting in undefined behaviour.