Add new unit tests for multiple transmission run inputs
The code now supports multiple transmission runs in the same input box (for first or second transmission run). Add new unit tests for these.
Also fix a previous test which expected that a comma in the transmission input was invalid - this is no longer invalid so replace it with alternative invalid text in the test.