Avoid extra prefix in loaded workspace names
If the instrument name is specified in the run number, e.g. INTER13460, and the workspace needs to be loaded, then it will now be loaded with a name excluding the instrument name, like TOF_13460, rather than including the instrument name, i.e. TOF_INTER13460. If the workspace is already in the ADS, then the existing workspace name will be used, i.e. INTER13460, without any prefix being added. A new unit test has been added for the latter case.

Re #24603