Set up postprocessing algorithm for groups
Add functions to set up the stitching algorithm and set properties based on the reduction configuration.
Pass a set of rows or a group to the batch algorithm runner at a time. When the batch completes, check for the next group/set of rows that needs to be processed and repeat until all are done

Re #24804