Add a flag to skip failed rows in the current round of processing
When we want to reprocess failed items, we only want to have one attempt at reprocessing them, so we set a skipped flag to make sure they don't get attempted again. This gets reset when starting a new round of processing. It will also get reset by autoprocessing when this is implemented when it automatically starts a new round of processing.

Re #24804