Fix batch presenter unit tests
Due to recent changes, the batch presenter will now only perform processing if the job runner returns a non-empty list of algorithms, so make sure that the mock returns something.
Add a new unit test for the case where the next batch is started when there are still algorithms left to run after a previous batch has completed.

Re #24804