Add force_save option to interpreter.save, re #24812
If the user is being prompted to save files (option in the settings)
then the prompt has the final decision whether to save or discard.

If the user is NOT being prompted, then pressing File>Save Script or
CTRL+S will force save, this has 2 outcomes:
- Code editor contents already exist in a file, in which case the
file is just saved
- Code editor is a `New Tab`, in which case the user is prompted to select
where to save.
- If a code editor is closed via `X` or middle mouse click, the file is
just discarded without notice.