Cache Beamline objects when read for first time by ExperimentInfo
The standard LoadInstrumet caches the Beamline info when created but
this was missing from the ExperimentInfo version and subsequently
in the instrument load from processed nexus files. Fixing this
reduces the size of a WISH workspace by 200MB when loaded from
a NexusProcessed file.
This is a minimal fix for the 4.0.0 release. The master will branch
will have the LoadInstrument code refactored to avoid this duplication.