A more working progress bar, re #25063
Now uses a single ProgressObserver that oversees all events. This makes
the event triggering synchronous - no two events can trigger at the same time.

Handling an event requires changes in the GUI thread. Whenever an event
triggers, a QueuedConnection signal is emitted to the GUI. This queues
an update that will be done at some point in the future event loops.