Crop the IvsQ output of RROA if Q limits specified
If the momentum transfer min and/or max are specified, crop the IvsQ output workspace to these limits. This is an interim output workspace which is used further down the reduction for stitching, and it is important to crop unwanted regions before the stitch. The main output (IvsQ_binned) is already cropped as a result of the rebin step.

Re #23817