Show the log value if they are all the same
Since the SNS DAS puts a time-series for even then unchanging string
logs, see if they are all the same and just put it in the value column.
This was not made specific for strings because it was discovered that we
have other logs that have multiple "entries", but are all just one
value.