No longer clears the tracked algorithm on any Algorithm end
Previously the AlgorithmProgressPresenter would clear the tracked
AlgorithmID pointer on any Algorithm ending, this meant that child
algorithms would clear their parent's pointer.

This change adds a check to see if the algorithm ending is
the one being tracked, this way child algorithms that are not
tracked, do not clear anything.

Re #25509