Abins: support atom-resolved spectra
The "Atoms" option for the Abins algorithm currently allows a set of
elements to be selected, in which case the returned workspaces will
consist of spectra summed over all atoms of this element (divided by
isotope if appropriate).

This option is here extended to also allow entries of the form
{PREFIX}{I} where the {PREFIX} "atom_" is set by
AbsinsModules.AbinsConstants.ATOM_PREFIX and the index {I} is the
1-indexed location of the atom in the Abins SData object, which in
turn should reflect the position in the original ab initio calculation
parameters. This allows the contributions of inequivalent sites to be
compared even if they are the same isotope.