Avoid overwriting row and group state when autoprocessing
Transferring a duplicate row will not overwrite the original row (note that duplicate here only concerns the run numbers and angle; we assume if the user has changed other values they want to keep their changes so we do not overwrite when transferring new runs. Note that the reason we check the run numbers is that subsequent runs may come in for the same angle so we need to add them to the row and re-reduce).

Transferring will also not reset state for the original group as long as none of the rows have changed.

Re #23027