Update release notes with progress bar changes
The progress bar is now reporting progress more accurately so this needs to be mentioned in the release notes. Previously progress used to be reset when you paused and restarted processing. Now it always reports based on the number of completed rows out of the current selection (or whole table if there is no selection).

Re #23027