Fix a bug with incorrect notification of summation type change
When we restore defaults we update the view from the model. Notifications are disabled so that the view does not notify back to the presenter that settings have been changed. However there was a bug where onSummationTypeChanged was not being disconnected which is now fixed. This commit also tidies up the way slots are connected so that onSummationTypeChanged is connected/disconnected at the same time as the other slots. It also ensures all slots are disconnected for signals rather than specific named slots.

Re #23051