Fix a bug with cancelling when search fails
If the search login fails when starting autorprocessing there is a bug where the operation is only partially cancelled. Cancelling was not being done completely because we subsequently get a autoreduction-started notification. For now just allow autoprocessing to start and let the user cancel it so we are in a known state.

Re #23027