Move search results model into catalog searcher
This simplifies the logic in the runs presenter and also the ownership of the search results model. This is now owned by the runs view and a reference is obtained and managed by the catalog searcher. The model no longer needs to be reset because it is simply cleared when its contents are invalidated.
This also fixes a problem where notification was not being passed to the runs presenter when the search is complete if no results were found.

Re #23027