Add request notification when inserting rows/groups
Rather than the job tree view adding the row/group straight away, send a notification to the subscriber and leave it up to the subscriber to call back to insert the rows/groups. This means that for the reflectometry gui we can avoid editing the table if processing is in progress.

Re #23027