Fix a bug auto-saving group workspace output
If a row has a group workspace as input then it will also have a group workspace as output. If we enable auto-saving, the save algorithm cannot handle a group workspace as input so it throws an exception which is unhandled. This fixes the problem by saving all the child workspaces individually.

Re #23027