Fix bug with row highlighting
Row styling was not being applied when algorithms complete.
Some unnecessary calls to update the styling were done on batch
complete.
A call has been added to update styling when the batch starts - this is
because the styling is being cleared for some reason and this is a quick
workaround to fix it.

Re #23027