Fix Ubuntu 16 error when toggling normalization
... after adding errorbars to a figure.

This was caused by the list of artists inside the old and new
ErrorbarContainer being flattened, then looped over to udpate
properties. The error was raised when the list of artists were of
different length, e.g. if errorbar caps were added, they would try and
update from the bars.