Fix a bug pasting empty group onto empty group
This was caused by the original group being removed to leave an empty table in the job tree view before the new group is added. The job tree view adds a new empty group to avoid an empty table so then the pasted group ends up in an invalid location.

Re #26025