Abins 2D: GroupWorkspace created correctly for multiple q-order calc
Currently the outputs display quite nicely in the MantidPlot spectral
viewer but don't seem to work at all with workbench slice viewer.

Performance is also slow relative to the 3.9.2 code, in part due to
the new instability detection/threshold logic which spends a lot of
iterations finding an appropriate threshold.