Update GUI with new polarization option
Remove the old polarization correction inputs from the GUI and replace them with a simple checkbox which is equivalent to the old option of 'ParameterFile' if checked or 'None' if not checked. The 'PA' and 'PNR' options have been removed entirely. Note that although the PolarizationCorrections model could now be replaced with a simple bool, it has been left because we may extend it with additional options in the future.

Re #23818