Tell fit browser when to normalise on Workbench
The C++ FitPropertyBrowsers's attribute m_shouldBeNormalised was not
being set in Workbench so fit results were always based on un-normalised
data regardless of what is plotted.

This commit exposes the Fit button's QMenu to Python and uses the event
of the menu opening to set m_shouldBeNormalised. This event was used
since otherwise it would need to be set in 4 places: on opening the fit
browser, on changing workspace, on changing workspace index and on
changing the normalisation of the plot.