Fix issue with normalisation of fit plot guesses
When plotting a fit guess on a normalised line the plot guess is working
on normalised data, as it should. When actually plotting the line the
plot guess is then normalised again. This commit makes sure the fit
guess is never normalised and prevents plotting the guess from updating
axes labels.