Fix typos causing incorrect notification calls
This commit fixes a couple of incorrect function calls but note that this doesn't actually affect the functionality at the moment because at the moment both of these notifications just cause the widget state to be updated. This commit fixes the new unit tests that have been added, though.

Re #26498