Fix inter-test dependency problem
There is a problem resetting the state between tests when running the whole suite. A workspace is created for the test REFL instrument, and even though a new workspace is used each time, it picks up settings from a previously-loaded parameters file, if those settings are not overridden in the current parameters file being loaded (which they will not be if we're trying to use an empty/minimal parameters file). I'm not sure how to get around this other than to use a different instrument name - this commit adds a fix to use instruments REFLEMPTY and REFLMANDATORY for the empty/minimal parameters files required for the new tests. It would be better to come up with a more robust solution in future though, especially as we now have 3 copies of the test instrument definition.

Re #26530