Check default peak in registered peaks in getter
Crashes were caused on Workbench when the default peak was not in the
list of registered peaks and the user was unable to change the default
peak.

Check that the default peak is registered when retrieving it and default
to a Gaussian if it does not exist. Display a warning to the user if you
change the default.

re #26580