Fix failing plotscriptgenerator test
The failing test was comparing a sample script with the output from
"generate_script". It was failing because there were some float values
outputted in the script and floating point errors were causing a
mismatch.

Make this test more robust by mocking out the functions called in
"generate_script" so they always output the same thing.