Abins: remove redundant check for ab initio program
This is always going to be caught by the next step which compares
hashes of the files. It's very rarely going to be relevant because most ab
initio programs use different filename conventions. If it is
relevant (i.e. a STDOUT stored to *.out),
it's still not going to save a lot of time (relative to a typical
Abins run) by avoiding the SHA hash.

It's only a few lines, but a lot of Abins LOC deal with caching and
this should be simplified to improve maintainability of the code.