Reversing cosmetic change to AbinsParameters; it breaks four tests
The Abins cache system stores a hash of the AbinsParameters.py file
and compares this with the current file when loading. This will check
for differences in instrument parameters, which seems sensible. It
will also be triggered by formatting changes and parallelism options
that have no bearing on the results. It also means that every time
this file is changed, the content of four test data
hdf5 files needs to be updated or Mantid will fail unit testing.

It's also not clear if this process will work as intended if the user
edits AbinsParameters via the Python API rather than the file. Editing
variables through the API is the appropriate method for e.g. parameter
sweeps so a caching method based on file checks may inappropriately
load stale data.

This is not a good system, "advanced parameters" should be managed as
a dictionary. This will also permit such parameters to be set from the
CLI.