Fix plot options hide curve on error bars bug
When setting the tick box to hide an error bar plot in the plot
options dialog, after closing and reopening, the "Hide curve" tick box
was no longer ticked. This was caused by toggling the visibility of the
error bars after re-plotting. The hide error bars check box was not
ticked so the error bars were being set to visible.