Abins caching: instead of file hash, store JSON of AbinsParameters
Advanced Abins parameters (i.e. parameters set as quasi-global
variable instead of through algorithm options) are validated against
the data cache to avoid re-loading stale data. However, the existing
implementation for this validation is to compare a hash of the
AbinsParameters.py file. This is needlessly fragile against formatting
changes, and insensitive to changes made through the Python API
after the module has been loaded.

The new implementation stores a JSON dump of parameters, and performs
comparison between Python dictionaries (so should be insensitive to
sorting and formatting.) As a bonus, this allows performance-related
parameters to be excluded from the validation process so one may
e.g. change parallelism settings without invalidating the cached data.

The resulting string is only a few times as long as the hash, and much
more informative. Hashing is still used to validate the input ab
initio data as it is very unlikely that two different code outputs
would contain the same meaningful data.