Abins: update CRYSTAL test for new parameters data format
The AbinsParameters.py file hash has been replaced with a dictionary
dump. The input format tests need to be updated to reflect this; the
comparison has been changed to a dict comparison and the sample data
for CRYSTAL calculations has been updated. MD5 files are updated here
in the main source; the corresponding data files have been uploaded to
the Mantid data repository.