Improve check on post-processing properties
Convert the result to a map so we can check the specific properties
we're interested in, rather than having to match the whole string.

To do this, extend the parsing functions so that different separators
can be used rather than just a comma. Add new unit tests for the new
parsing options.

Re #26512