Create 1:1 spec-det mapp when message is missing
When subscription to the topic that provides the spectrum to detector ID
mapping fails or the message is otherwise not consumed generate a 1:1
mapping by loading the instrument before creating the buffers.

This is a slightly hacky solution in time for V20 beam time, the
spectrum to detector ID mapping may be later moved to the run start
message, making this easier to handle.

Fixes #26692