Document BufferThreshold for Kafka event stream
Gives a brief overview of what this parameter does and roughly what it
should be set to.

This may not be the most optimal place for this documentation, but is
likely the first place users will look given that this is where you get
to from the Load > Live Data help button.

See https://jira.esss.lu.se/browse/DAM-1353