File production seems to be OK
Main change here is preventing LoadNexusProcessed from running the
instrument save when going via the ESS route.

New problem is that LoadNexusProcessed does not attempt to look for new
Geometry in Nexus as part of the loading, and therefore gives up when it
shouldn't

LoadNexusProcessed also annoyinly looking for group named "instrument",
which we now do not adhere to.