Use try-finally block when saving/loading project
When the _save funtion is called to save a project, the __is_saving
variable is set to True. This prevents the user from closing Workbench
whilst the saving is going on. The __is_saving variable is then set back
to False at the end of the function, however if the function exits early
then the variable will still be True. The try-finally block ensures the
variable is reset. Similar for loading.