Re #25686 Fix cost functions failing on GSL 2.3 with no params
GSL 2.4 removed checks on vector and matrix views that enforced n>=1
elements. When allocating views with no elements (from no params) RHEL7
would throw an abort.

Whilst this is an unlikely edge case, we still need to handle it. When
we encounter a n=0 length view or vector we allocate a single element
(or two for a matrix) with value 0.