Add "filterMessages" method to MessageDisplay
This method can filter messages shown in the display by whether or not
they origintated from the framework or a Python script.

Store message history in a QList. We can then recall from this list when
we are filtering.

Separate context menu generation from context menu "showing". This will
allow us to add items to the menu in Python and Workbench.