Do not save "Show All Script Output" setting
There's not much point in storing the setting. When a user executes a
script that was not executed before in that session, the output for that
script is automatically enabled. This in turn will automatically check
the "showAllScriptOuput" option since all script output is enabled.
Hench the setting would always be overwritten.