Abins: Remove iterative thresholding process
The semi-empirical powder averaging procedure of Abins includes a lot
of logic dealing with "stability". This was introduced as part of
https://github.com/mantidproject/mantid/pull/19906 with the stated
goal

> to avoid non-physically high intensities of higher orders quantum
> events.

It is related to the process of removing small values of S with a
numerical threshold. The thresholding process is a useful time-saver
as it cuts down the N of the O(N^4) enumeration of high-order quantum
events. The additional "stability" procedure employed here is to check
if the order N excitations have a higher intensity than order N-1
excitations, and if so to recompute all the excitations using a
_higher_ threshold.

This code is causing several problems and I propose in the short term
to deal with them by simply removing this functionality and
streamlining the module.

- The current maintainer/developer of Abins (myself) does not
  understand how the code is supposed to work. Why does raising the
  threshold help in this situation? How many iterations are typical?
  Without this knowledge the feature cannot safely be adapted to deal
  with new situations.
- The comments within the code generally refer to this procedure as a
  matter of numerical stability; this is quite confusing when it
  appears that the physics is being checked rather than numerical
  convergence.
- It's not clear that removing low-intensity contributions will result
  in better physics, even if it does fix a surprising trend in
  intensity.
- The implementation relies heavily on state; tracking previous values
  of thresholds via object attributes as loops are iterated, and using
  methods to modify these attributes. This makes it difficult to
  change the flow of the program, impeding work on features and
  optimisation.

It seems simpler, faster and more useful to perform checks at the end
of the Abins run and suggest to the user that they may wish to run
again with a different threshold.