Abins: add a check for small S threshold parameter
A method is added to SData allowing it to compare the magnitude of S
values with the thresholds set in AbinsParameters. If the maximum
value of S multipled by s_relative_threshold (1% by default) is less
than s_absolute_threshold, this suggests that some significant
excitations are being removed and a warning is raised. The method is
called by SPowderSemiEmpiricalCalculator after all the S data has been
computed or read from the cache.

This should provide useful feedback to the user about the thresholding
process, without automatically performing any potentially-unstable
threshold adjustments.

A limitation in the approach here is that the check is being performed
_after_ the thresholding was applied, so would miss a case in which
_all_ the data is removed. In the current implementation of
thresholding, there will always be at least two points left, so the
warning would still be raised. Other implementations may need to
perform this check at an earlier stage; it does seem logical to
attaching this method to SData. A call to .flatten() ensures that the
method is compatible with multi-dimensional arrays of S values.