Abins documentation: update "concepts" documentation
The main documentation for the Abins algorithm focuses on the user
interface and how to run Abins.
A more detailed overview of the approximations employed in
Abins (DynamicalStructureFactorFromAbInitio.rst) is
stored in the "concepts" section of the docs and linked to by the
Algorithm docs.

This commit mostly tweaks the wording for improved clarity. An
explicit mention is also added of the current restriction to
Gamma-point data.