Make jedi call tips persistent in completion api
QSci has a bug where it will not display call tips in the GUI the first
time you enter a bracket for a function. You need to either enter an
argument and a comma, or delete the bracket and put it back in again.
Since jedi's completions are updated on every key movement the call tips
will never show unless they are persistent.