Fix errors when not giving specNum to plot func
Plotting a spectrum with one histogram without giving a specNum was
possible, however things like fitting and plot options needed a
spectrum to operate. Assign spec_num=1 to the WorkspaceArtist in these
cases.

Remove incorrect tests from helperfunctionTest

The tests were testing the behaviour of get_wksp_index_and_spec_num when
it was not passed either a wkspIndex or specNum. The code dealing with
this case has been pulled back into MantidAxes.plot so that the
calculated specNum can be set on the WorkspaceArtist. Hence the removed
tests are no longer needed.