Fix MD Q conversion for indirect spectroscopy
The naming of variables such as ki, ei & e_tr caused confusion
when calculating Q3D in the Indirect case. The original code
set the magnitude of the incident K-vector to m_ki but for
an indirect instrument this is actually the magnitude of the
final K vector. The fix includes better naming to clarify
the intent of each variable.

Refs #26965