Fix bug on home button press with markers on plot
When the home button is pressed, the view limts are updated to centre
on the current content. When markers are present, their limits are
updated to span the new limits. This means that every time home is
pressed the view will zoom out to give the markers a margin.

Temporarily hiding the markers whilst the new limits are set excludes
the markers from the limit calculation.