Use import aliases in auto completions
Given an import "import numpy as np" the call tips were previously
stored as e.g. "numpy.arange". This displayed "arange (numpy)" as a call
tip but would not display the parameters after completing on "np.ar".
We now store the completion as "np.arange" so that parameters are
displayed after the completion.