Use focusWindowChanged to keep figures on top
Hook into QtGuiApplication's focusWindowChanged signal to raise
FigureWindows when Workbench's MainWindow gets focus.

This does produce a brief flicker as the FigureWindow is hidden and
brought back to the front.

The behaviour we are after is:
    - The FigureWindow should be on top of the MainWindow if the
      MainWindow has focus.
    - You should be able to minimize the FigureWindow.

This was the best multi-platform solution I could find. Setting the
FigureWindow as a child of MainWindow kept the FigureWindow on top only
on Windows. Making the FigureWindow a Qt::Tool meant the window could
not be minimized, so would perpetually block the main window. Making the
figure an MdiSubWindow produced a similar issue, with the added quirk
that maximizing the FigureWindow would replace MainWindow's contents,
with no way of getting it back.

There was some promise using the Qt::WindowAlwaysOnTop flag, toggling
the flag on and off when focusWindowChanged signals were set off. I ran
into some problems on Ubuntu where grabbing the title bar of the
FigureWindow was indistinguishable from clicking an external window.