Set MainWindow as FigureWindow's parent on Windows
The current method of keeping the FigureWindow on top of Workbench's
MainWindow is to hook onto the foucsObjectChanged event and raise the
FigureWindow whenever the MainWindow gets focus. This causes a flicker
as the FigureWindow is breifly hidden before being restored.

On Windows, setting the MainWindow as the parent of our FigureWindows
will always keep the FigureWindow on top, without any flicker. However
this only seems to be the case on Windows. On Ubuntu, we still need to
manually raise the FigureWindow.