Push zoom state to stack after home button pressed
Pressing the home button did not previously add the zoom state of the
figure to the stack. This meant that zooming in (to state 1) then
pressing home (state 2), zooming in again (to state 3), then pressing
back, would get you back to state 1; skipping state 2.

Remove redundant comment and unused signal.