Fix subtle variable lifetime bug when using muParser
DefineVar takes a pointer to a location of a variable so this memory
address must be valid for the duration of the Parser objects use. The
try/catch limited the scope of the temporary variables to the try{} block
and accessing them later would lead to undefined behaviour. It took
clang 10 to pick this up. Use DefineConst to define constants instead.