Use IObject::distance to compute virtual peak position
For some peaks not on a detector the intersection code
can given false positives that not intersection is
possible due to precision issues. In our case we only
need the distance to any intersection point on the surface
regardless of whether it is an entry/exit/undefined type
of track.
Refs #26690