Do not plot errors in fig interaction test on RHEL
The old version of Matplotlib on RHEL7 was throwing an error when trying
to draw an interacive legend in a test. An error was thrown when trying
to retrive the "button_release_event" from the canvas, in order to link
it to the legend.

The test still fulfills its purpose wihout the plots being errorbar
plots, so turn off errorbars on the old version of Matplotlib that RHEL7
uses.