Remove spurious usage reports for populating workspace list
On the Save tab, avoid calling the slot to populate workspaces lists
when the presenter subscribes to the view. This causes spurious entries
in the usage reports every time the GUI is opened.

The logic for when the workspaces list gets populated (as well as when
the default save path is set) should actually be in the presenter
anyway. This commit moves it there and adds some tests for it.

Re #27030