Fix a bug where the GUI could not be closed
This commit fixes a couple of logic errors in handling the
processing/autoprocessing state which was causing the GUI to be
considered unclosable under certain situations, which in turn meant that
MantidPlot/Workbench could not be closed.

The specific problems were:
- A bad logic check on whether any batch is processing or
autoprocessing.
- Incorrectly setting the isProcessing flag when requesting to resume
Autoprocessing. This has been removed so that it now only gets set when
(and if) processing genuinely happens.

Re #26691