Fix parent algorithm tests
ReflectometryReductionOne tests are updated to set a run number on the
input workspace, because this is used when creating default transmission
workspace names and this allows us to add tests that the correct
transmission outputs are created. Set a run number by default because
this is more widely useful. There are a couple of tests that need it not
to be set so they clone the workspace and remove it.

Re #26896