Set the output workspace even if a name is not found
This reverts to previous behaviour where, when run as a child, the
parent algorithm gets the output workspace even if it doesn't have a
name set. A clause has been added so that we don't try to add it to the
ADS if it doesn't have a name.

A separate issue has also been fixed to identify the case where the
second transmission run number is not found and to skip setting the name
in this case. Previously, the name would be set to include only the
first transmission run number, which is misleading.

Re #26896