Revert "Add 'on_top_of_main_window' decorator to workbench"
This reverts commit 1c948d12fc09ee33f96d23c1d467c47d94fc0cf9.

After moving the functionality for the plot window staying on top of the
main window to a decorator, having multiple plots open was causing some
major errors, see #27170.

Since the 'on_top_of_main_window' decorator is not used by anything else
yet and we are close to a release, I suggest we revert to having the
functionality sit in the FigureWindow. We should raise an issue to make
the functionality more portable, which can be looked at after the
release.