Create symlinks only for existing targets in macOS package
The Resources and Helpers directories in the frameworks
are optional and symlinks should not exist if the target
does not exist. Whilst having the symlinks doesn't affect
the functionality of the program it creates issues using the
cp -r command as we do in systemtests as it exits with a
non-zero status when trying to following dead links.