Set the correct facility for the reflectometry GUI
The reflectometry GUI sets the default instrument in the config service
when it starts up to one of the ISIS reflectometers. This commit adds a
check that the facility is ISIS and if not it sets the facility to ISIS.
This avoids confusion where you essentially have invalid config. It also
avoids confusion where e.g. loading runs from the default archive might
not work because the archive may be a different facility.

I have also added code to save the new config settings to file. Again
this is to avoid confusion between what users see in mantid and what is
in the saved settings. It is particularly confusing otherwise because
Workbench will write settings to file when it closes but MantidPlot will not.

Re #27241