Set the correct facility for the SANS GUI
The ISIS SANS GUI sets the default instrument in the config service
to one of the ISIS SANS instruments. This commit adds a check that the
facility is ISIS and if not it sets the facility to ISIS. This avoids confusion
where you essentially have invalid config. It also avoids confusion where e.g.
loading runs from the default archive might not work because the archive may be
a different facility.

Re #27241