Re #27151 Check if colour fill bar is present before removing
Checks if the colour fill bar is present before attempting to remove it.
This fixes numerous unit tests, and (I suspect) prevents us throwing if
a user has tampered with their colour fill bar.

Additionally, there is a change for the minimum plot value to reflect
the changes where we shouldn't be attempting to scale to log(0)