Corrected requested changes incl. equation for exp parameters
(1)function estimateExpParameters now does not take any input, but instead uses
m_numberofExpononetials inside the function (and in the if statement of
setNumberOfExponentials). (2) Corrected equation to get lifetime of first
exponential from two data points (3) Corrected typo in comments regarding choice of
second exp parameters from 20% to 10% and removed comment on accounting for
background. (4) variables for height and lifetime of second exponential are not
declared inside the for-loop over the spectrum number.