Re #27297 Preserve custom sample thickness
A users sample thickness would be changed on loading or processing a
file. This checks if there is a custom sample thickness of any sort then
only takes the file one if there is not one present.

Additionally, we warn if we cannot convert a sample thickness in a more
specific way, and fall back to the file thickness in this case