RE #27158 Fix crash when setting max cores
If the maximum number of cores mantid can use is set to be lower than
the number of cores on the machine, certain algorithms would ignore this
limit and run extra threads.
As a result, workspaces that were not created with this higher number of
threads would cause a seg fault when the algorithm attempted to access
an index of a vector that was out of range.
Manually setting the number of threads allowed back to the value in the
properties file is a solution to the crash but has not fixed the root
cause of the problem.