Change the IvsQ output to be scaled
Previously the ScaleFactor was only applied to the IvsQ_binned output.
This commit also ensures that it is applied to the IvsQ output. This is
more in keeping with other settings, e.g. cropping to the Qmin/Qmax is
also applied to the IvsQ output. This is important because this output
is used when stitching results together so this ensures that the
stitched results also obey the scale and cropping specified.

Re #27317