Renames fit workspaces when multifit but throws exception for one run
Code that is commented out will rename the fit workspaces from the naming of the results workspaces when spectra from more than one run are fitted . However when fitting multiple spectra from a single run an exception is thrown relating to read access violation L644 of shared_count.hpp. The file changed in this commit does appear in the call stack but it is unclear how the changes cause this exception.