Fix problem with naming of output workspaces from Jump Fit
The Jump Fit can takes as its input the output from the ConvFit tab. The Jump
Fit input workspace name therefore often has the text _Result_ in the middle
of it. The function JumpFitModel::constructOutputName attempts to remove this
but was searching for "_Results" instead of "_Result" which had the unwanted side
effect of remove the "_Results" text from the end of the workspace name
(leaving the text _Result_ in place in the middle of the name).

I think this problem was introduced in pull request 24348 when some name changes
were made to the workspace groups that are created by the Jump Fit process