Incorporate review comments
I've modified getDockedOrFloatingWindow so it's more general and
works when the IndirectSettings window is an independent window
and also has a parent. I'm not convinced the #if #endif that checks
the Qt version (essentially whether we're in Workbench or MantidPlot)
is absolutely needed here but since it's late in the release I've
just done a minimal change

The 2nd review comment was about the Indirect, Settings window being modal
in Workbench but not in MantidPlot. The IndirectSettings widget in MantidPlot
isn't an independent window and is embedded two levels down below the parent
window. So before this change the windowModality setting in the IndirectSettings.ui
file wasn't doing anything when settings was accessed directly from the Indirect
menu (windowModality only meaningful for independent windows). So I have copied
the modality setting read from the .ui file up two levels and it now works.