Abins broadening - normalisation and tidying
A normalisation term was found in the Gaussian function which scales
using Sigma and fwhm to fix the maximum value of the peak as sigma is
varied. This does not appear to be consistent with the function
defined in the Abins paper, and leads to an increasing weight for
high-energy terms. This term is removed.

Instead a normalisation feature is added which normalises the sum of
the output values to one. This is required for a convolution kernel
that does not change the overall intensity, and removes any dependence
on the bin width. The cost of this process is not insignificant; it
may be more efficient to use the analytic integral ("normal" function)
which is 'pre-normalised'. Testing with different bin widths finds
that the differences in the resulting curves are negligible at typical
bin width and with large bins the results, while different, are not
obviously superior or inferior.

Work is started on refactoring the broadening methods that use
truncated functions by providing a truncation function that accepts
the `gaussian` or `normal` function as an input.